requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pymysql
pandas
sqlalchemy
cryptography
python-dotenv
